requirements.txt
pandas==2.1.4
numpy==1.26.2
sqlalchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
scholarly==1.7.11
arxiv==2.1.0
networkx==3.2.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
streamlit==1.29.0
anthropic==0.7.8
python-dotenv==1.0.0
tqdm==4.66.1